Move get_last_change_idx off the runtime path, and make that a rule
Precomputing the stop index into the shortcut binary removed the last
per-query call to get_last_change_idx, but the function stayed in
timezonefinder/utils_numba.py - where numba compiles and binds it at import,
in every user's process, for something no lookup runs. Dead on the path it was
written for, and not free.

It is now plain Python in shortcut_index.py beside the builder that writes
those values, which is its only caller. data_integrity re-checks the committed
values against that same implementation, so the two still cannot drift, and its
unit test moved with it.

An audit of everything this branch added found no other dead symbols: no
unreferenced module-level names and no write-only constants across
shortcut_index.py, layout.py, scripts/utils_numba.py and the new prototype.

CLAUDE.md gains the rule, because the failure mode is not noticing rather than
not caring: removing the last caller of something is half the change, and the
other half is deciding what the callee now is. No callers means delete; callers
only in tests means say so; callers only at build time means leave the runtime
modules.
