+++ CHANGELOG.rst
@@ -17,6 +17,8 @@ X.X.X (unreleased)
 
 Internal:
 
+* ``get_last_change_idx`` moves out of the runtime modules to sit beside the shortcut index builder, its only caller. Precomputing the value into the shortcut binary removed the last per-query call to it, but the ``@njit`` function stayed in ``timezonefinder/utils_numba.py``, where numba compiles and binds it at import in every process for something no lookup runs. It is now plain Python next to the builder that writes those values, and ``scripts/data_integrity.py`` re-checks the committed values against that same implementation, so the two still cannot drift
+
 * the shortcut lookup's slot arithmetic is one shift and one mask rather than two shifts, two masks, a multiply and an add. H3 stores a cell's base cell immediately above its digits, so the expression the index was addressed by is a single contiguous bit field - an identity over any 64-bit value, not a property of the cells that exist, and pinned by a test that asserts it over arbitrary integers as well as real ones. Worth ~5% of a unique-zone query, measured paired and order-alternated over 61 rounds. Accelerating the same arithmetic with numba or the C extension was measured and refused: an empty numba call alone costs about two thirds of the entire stage, so no kernel can pay for the dispatch - recorded in ``potential-improvements.md``, since it rules out the whole family rather than one attempt
 
 * a new prototype, ``prototypes/shortcut_resolution_query_bench.py``, swaps a resolution 4 shortcut index underneath the real ``TimezoneFinder.timezone_at`` to settle whether the finer resolution pays for itself on a whole query, paired and order-alternated as this repository's methodology requires. Its timing comparison never ran: the correctness gate that precedes it found the two indices answering one fixture point differently, and resolution 4 is the wrong one there. The cause is pre-existing and unrelated to the index format - the shortcut compiler decides polygon/cell overlap by vertex inclusion alone and never tests whether a polygon edge crosses the cell, a simplification whose stated precondition (that polygons and cells have a similar size) degrades as the resolution rises. Recorded in ``potential-improvements.md`` as a blocker on any resolution change, together with a second, narrower gap it turned up at the poles, where cells above latitude 88 can omit the polygon covering them. Both are invisible to a realistic workload today: sampled by area rather than uniformly in latitude, 3,000 points produce no wrong answer at either resolution. Nothing in the package changes
